requirements.txt
anyio==3.3.0
appdirs==1.4.4
beautifulsoup4==4.9.3
bibtexparser==1.2.0
black==21.7b0
bs4==0.0.1
cached-property==1.5.2
certifi==2021.5.30
chardet==3.0.4
click==8.0.1
colorama==0.4.4
commonmark==0.9.1
dictdiffer==0.8.1
dominate==2.5.2
feedparser==5.2.1
future==0.18.2
h11==0.12.0
httpcore==0.13.6
httpx==0.18.2
idna==2.10
mistletoe==0.7.2
mypy-extensions==0.4.3
notion-client==0.7.1
numpy==1.21.1
pandas==1.2.1
pathlib==1.0.1
pathspec==0.9.0
Pygments==2.9.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-slugify==4.0.1
pytz==2021.1
Pyzotero==1.4.20
regex==2021.8.3
requests==2.24.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.2.1
text-unidecode==1.3
tomli==1.2.1
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.25.11
